test & vars name fixes
